requirements.txt
python-telegram-bot>=20.0
httpx>=0.27.0
pymysql>=1.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
